pyproject.toml
[tool.poetry]
name = "llm-gateway"
version = "0.1.0"
description = "Service to facilitate access to large language models"
authors = ["Data Science & Engineering at Wealthsimple <data@wealthsimple.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.95.0"
pydantic = "^1.10.7"
psycopg2-binary = "~2.9.3"
alembic = "~1.8.0"
sqlalchemy = ">=1.3.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
openai = {extras = ["datalib"], version = "^0.27.4"}
cohere = "^4.6.1"
starlette-exporter = "^0.16.0"
fastapi-route-logger-middleware = "^0.2.3"
opentelemetry-api = "^1.19.0"
opentelemetry-exporter-jaeger-thrift = "^1.19.0"
opentelemetry-propagator-jaeger = "^1.19.0"
opentelemetry-sdk = "^1.19.0"
opentelemetry-instrumentation-fastapi = "^0.40b0"
opentelemetry-instrumentation-logging = "^0.40b0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
black = "23.3.0"
isort = "5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.2"
pytest = "^7.3.0"
urllib3 = "1.26.15"  # https://github.com/psf/requests/issues/6437

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
